Fix PR review comments: move test to appropriate file and remove unnecessary mock
